requirements.txt
accelerate==0.21.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
anyio==3.7.1
assemblyai==0.17.0
asttokens
async-timeout==4.0.3
attrs==23.1.0
backcall
backports.functools-lru-cache
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama
comm
comtypes==1.2.0
contourpy==1.1.0
cycler==0.11.0
dataclasses-json==0.5.14
debugpy
decorator
diffusers==0.20.0
exceptiongroup==1.1.3
executing
fastapi==0.103.1
ffmpy==0.3.1
filelock==3.12.2
fonttools==4.42.0
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
gradio==3.42.0
gradio_client==0.5.0
greenlet==2.0.2
gTTS==2.3.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
importlib-metadata
importlib-resources==6.0.1
ipykernel
ipython
jaraco.context==4.3.0
jedi
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client
jupyter_core
keyboard==0.13.5
kiwisolver==1.4.4
langchain==0.0.273
langsmith==0.0.24
llvmlite==0.40.1
lxml==4.9.3
MarkupSafe==2.1.2
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline
more-itertools==10.1.0
mpmath==1.2.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio
networkx==3.0
numba==0.57.1
numexpr==2.8.5
numpy==1.24.4
openai==0.28.0
openai-whisper
openapi-schema-pydantic==1.2.4
opencv-python==4.8.0.76
orjson==3.9.5
packaging
pandas==2.0.3
parso
pickleshare
Pillow==10.0.0
platformdirs
playwright==1.37.0
prompt-toolkit
psutil
pure-eval
py-cpuinfo==9.0.0
PyAudio==0.2.13
pycparser==2.21
pydantic==2.2.0
pydantic_core==2.6.0
pydub==0.25.1
pyee==9.0.4
pygame==2.5.1
Pygments
pyparsing==3.0.9
pypiwin32==223
pypng==0.20220715.0
python-dateutil
python-multipart==0.0.6
pyttsx3==2.90
pytube==15.0.0
pytz==2023.3
pywin32==305.1
PyYAML==6.0.1
pyzmq
qrcode==7.4.2
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
rpds-py==0.10.2
safetensors==0.3.2
scipy==1.11.2
seaborn==0.12.2
semantic-version==2.10.0
six
smmap==5.0.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.4.1
SpeechRecognition==3.10.0
SQLAlchemy==2.0.20
stack-data
starlette==0.27.0
sympy==1.11.1
tenacity==8.2.3
termcolor==2.3.0
tiktoken==0.3.3
tokenizers==0.13.3
toolz==0.12.0
tornado
tqdm==4.66.1
traitlets
transformers==4.31.0
typing-inspect==0.9.0
typing_extensions
tzdata==2023.3
ultralytics==8.0.157
urllib3==2.0.4
uvicorn==0.23.2
wcwidth
websockets==11.0.3
whisper==1.1.10
wikipedia==1.4.0
wolframalpha==5.0.0
xmltodict==0.13.0
yarl==1.9.2
youtube-transcript-api==0.6.1
zipp
